\name{add_subtree}
\alias{add_subtree}
\title{Add Nodes To a Plot of a Hierarchical Clustering Model}
\description{
  Used to add nodes to a plot region created for visualising hierarchical clustering models arising from functions mcdc, mddc and ncutdc. Not intended for use outside function tree_plot.
}
\keyword{internal}
