\name{breastcancer}
\docType{data}
\alias{breastcancer}
\title{Discrimination of Cancerous and Non-Cancerous Breast Masses}
\description{
  This data set contains case information from 699 patients from Wisconsin General Hospital who received examination for mammographic masses, which were classified as either benign or malignant.
}
\usage{breastcancer}
\format{A list with entries $x (a 699x9 matrix with each row corresponding to an individual case) and $c (a vector of labels indicating whether the mass was diagnosed as benign (2) or malignant (4)).}
\source{UCI Machine Learning Repository.}
\references{
  Lichman, M. (2013) UCI Machine Learning Repository. Irvine, CA: University of California, School of Information and Computer Science. [https://archive.ics.uci.edu/ml] 
}
\keyword{datasets}
