\name{dermatology}
\docType{data}
\alias{dermatology}
\title{Eryhemato-Squamous Disease Identification}
\description{
  This data set contains clinical and histopathological information from 366 cases of 6 different skin disorders/diseases for the purpous of differential diagnosis of eryhemato-squamous disease.
}
\usage{dermatology}
\format{A list with entries $x (a 366x34 matrix with each row corresponding to an individual case) and $c (a vector of labels indicating the skin condition).}
\source{UCI Machine Learning Repository.}
\references{
  Lichman, M. (2013) UCI Machine Learning Repository. Irvine, CA: University of California, School of Information and Computer Science. \url{https://archive.ics.uci.edu/ml}
}
\keyword{datasets}
