\name{hp_plot}
\alias{hp_plot}
\title{Visualise a Hyperplane Separator for Clustering}
\description{
  Provides a visualisation of the partition induced by a hyperplane separator generated by one of mch, mdh and ncuth.
}
\usage{
hp_plot(sol, labels)
}
\arguments{
  \item{sol}{a solution arising from one of the functions mch, mdh and ncuth.}
  \item{labels}{(optional) a vector of class labels. if provided then points in different classes are plotted in different colours. the external validity of the partition is also indicated when this can be computed using true class labels.}
}
\keyword{internal}
