\name{optidigits}
\docType{data}
\alias{optidigits}
\title{Optical Recognition of Handwritten Digits}
\description{
  This data set contains vectorised images of handwritten digits (0--9), compressed to 8x8 pixels.
}
\usage{optidigits}
\format{A list with entries $x (a 5620x64 matrix with each row corresponding to an image) and $c (a vector of labels indicating the written digit).}
\source{UCI Machine Learning Repository.}
\references{
  Lichman, M. (2013) UCI Machine Learning Repository. Irvine, CA: University of California, School of Information and Computer Science. \url{https://archive.ics.uci.edu/ml}
}
\keyword{datasets}
