\name{pendigits}
\docType{data}
\alias{pendigits}
\title{Pen-based Recognition of Handwritten Digits}
\description{
  This data set contains features derived from pen trajectories arising from handwritten digits (0--9) from 44 subjects.
}
\usage{optidigits}
\format{A list with entries $x (a 10992x16 matrix with each row corresponding to a pen trajectory) and $c (a vector of labels indicating the written digit).}
\source{UCI Machine Learning Repository.}
\references{
  Lichman, M. (2013) UCI Machine Learning Repository. Irvine, CA: University of California, School of Information and Computer Science. \url{https://archive.ics.uci.edu/ml}
}
\keyword{datasets}
