\name{phoneme}
\docType{data}
\alias{phoneme}
\title{Speech Recognition through Phoneme Identification}
\description{
  This data set contains log-periodogorams generated from sound frequency recordings of 5 spoken phonemes.
}
\usage{phoneme}
\format{A list with entries $x (a 4509x256 matrix with each row corresponding to a periodogram) and $y (a vector of labels indicating the spoken phoneme).}
\source{The Elements of Statistical Learning Theory.}
\references{
  Friedman, J., Hastie, T., Tibshirani, R. (2001) The Elements of Statistical Learning. \emph{Springer series in statistics.} \bold{1}, pp. 241--249 [https://web.stanford.edu/~hastie/ElemStatLearn/data.html]
}
\keyword{datasets}
