\name{ppclust.optim}
\alias{ppclust.optim}
\title{Optimisation Call for Projection Pursuit Algorithms}
\description{
  Provides a gateway to R's base optimisation function optim, with additional options as required by projection pursuit algorithms within the package. Not intended for independent use.
}
\keyword{internal}
