\name{plot.ppci_hyperplane_solution}
\alias{plot.ppci_hyperplane_solution}
\title{Visualise a Hyperplane Separator for Clustering}
\description{
  Provides a visualisation of the partition induced by a hyperplane separator generated by one of mch, mdh and ncuth.
}
\usage{
\method{plot}{ppci_hyperplane_solution}(x, labels = NULL, transparency = NULL, ...)
}
\arguments{
  \item{x}{a solution arising from one of the functions mch, mdh and ncuth.}
  \item{labels}{(optional) a vector of class labels. if provided then points in different classes are plotted in different colours.}
  \item{transparency}{(optional) if ommitted then points in the scatterplot are
  shown as solid. If set to a value in (0, 1) then points are shown with transparency. Lower values correspond with a greater degree of transparency.}
  \item{...}{additional graphical parameters. Currently none implemented.}
}
\seealso{mch, mdh, ncuth, hp_plot}
\value{None}
\keyword{internal}
